Create SudokuGUI.py
This is my Sudoku GUI ad it shows the gameplay, you write the number you want in the box as a guess, and press enter to place an answer then it will show if your answer is right or wrong. You can also press space so the program will solve the game for you.